+++ constructr-coordination-consul/src/main/scala/com.tecsisa.constructr.coordination.consul/ConsulCoordination.scala
@@ -19,15 +19,19 @@ package consul
 
 import java.nio.charset.StandardCharsets.UTF_8
 import java.util.Base64
-import java.util.Base64.{getUrlDecoder, getUrlEncoder}
+import java.util.Base64.{ getUrlDecoder, getUrlEncoder }
 
 import akka.Done
-import akka.actor.{ActorSystem, Address, AddressFromURIString}
-import akka.event.{Logging, LoggingAdapter}
+import akka.actor.{ ActorSystem, Address, AddressFromURIString }
+import akka.event.{ Logging, LoggingAdapter }
 import akka.http.scaladsl.Http
-import akka.http.scaladsl.client.RequestBuilding.{Get, Put}
+import akka.http.scaladsl.client.RequestBuilding.{ Get, Put }
 import akka.http.scaladsl.model.MediaTypes.`application/json`
-import akka.http.scaladsl.model.StatusCodes.{NotFound, OK}
+import akka.http.scaladsl.model.StatusCodes.{ NotFound, OK }
+import akka.http.scaladsl.model.headers.{
+  ModeledCustomHeader,
+  ModeledCustomHeaderCompanion
+}
 import akka.http.scaladsl.model.{
   HttpEntity,
   HttpRequest,
@@ -39,17 +43,57 @@ import akka.http.scaladsl.model.{
 }
 import akka.http.scaladsl.unmarshalling.Unmarshal
 import akka.stream.ActorMaterializer
-import akka.stream.scaladsl.{Sink, Source}
-
+import akka.stream.scaladsl.{ Sink, Source }
 import io.circe.Json
 import io.circe.parser.parse
-
 import de.heikoseeberger.constructr.coordination.Coordination
 
 import scala.concurrent.Future
-import scala.concurrent.duration.{Duration, FiniteDuration}
+import scala.concurrent.duration.{ Duration, FiniteDuration }
+import scala.util.{ Success, Try }
 
 object ConsulCoordination {
+
+  private final case class ConsulCoordinationSettings(
+      host: String,
+      port: Int,
+      agentName: String,
+      https: Boolean = false,
+      accessToken: Option[String] = None
+  )
+  private object ConsulCoordinationSettings {
+
+    def apply(actorSystem: ActorSystem): ConsulCoordinationSettings = {
+      val config = actorSystem.settings.config
+      val host   = config.getString("constructr.coordination.host")
+      val port   = config.getInt("constructr.coordination.port")
+      val agentName = Try(config.getString("constructr.consul.agent-name"))
+        .getOrElse("")
+      val https = Try(config.getBoolean("constructr.consul.https"))
+        .getOrElse(false)
+      val accessToken = Try(config.getString("constructr.consul.access-token")).toOption
+
+      ConsulCoordinationSettings(host, port, agentName, https, accessToken)
+    }
+
+  }
+
+  private final case class ConsulToken(value: String)
+      extends ModeledCustomHeader[ConsulToken] {
+    override def companion: ModeledCustomHeaderCompanion[ConsulToken] =
+      ConsulToken
+
+    override def renderInRequests: Boolean = true
+
+    override def renderInResponses: Boolean = false
+  }
+  private object ConsulToken
+      extends ModeledCustomHeaderCompanion[ConsulToken] {
+    override def name: String = "X-Consul-Token"
+
+    override def parse(value: String): Try[ConsulToken] =
+      Success(ConsulToken(value))
+  }
 
   final case class UnexpectedStatusCode(uri: Uri, statusCode: StatusCode)
       extends RuntimeException(
@@ -74,7 +118,7 @@ final class ConsulCoordination(
 
   @volatile var stateSession: Option[SessionId] = None
 
-  implicit val logger: LoggingAdapter = Logging.getLogger(system, this)
+  private implicit val logger: LoggingAdapter = Logging.getLogger(system, this)
 
   private val settings = ConsulCoordinationSettings(system)
   logger.debug("Initializing Consul Coordination using settings: {}", settings)
@@ -202,13 +246,15 @@ final class ConsulCoordination(
     send(Put(uri)).flatMap {
       case HttpResponse(OK, _, entity, _) => ignore(entity).map(_ => Done)
       case HttpResponse(NotFound, _, entity, _) =>
-        ignore(entity).flatMap { _ =>
-          logger.warning(
-            "Unable to refresh, session {} not found. Creating a new one",
-            sessionId
-          )
-          createIfNotExist(self, ttl)
-        }.map(_ => Done)
+        ignore(entity)
+          .flatMap { _ =>
+            logger.warning(
+              "Unable to refresh, session {} not found. Creating a new one",
+              sessionId
+            )
+            createIfNotExist(self, ttl)
+          }
+          .map(_ => Done)
       case HttpResponse(other, _, entity, _) =>
         ignore(entity).map(_ => throw UnexpectedStatusCode(uri, other))
     }
@@ -284,8 +330,9 @@ final class ConsulCoordination(
     send(Put(createSessionUri, sessionEntity)).flatMap {
       case HttpResponse(OK, _, entity, _) => unmarshalSessionId(entity)
       case HttpResponse(other, _, entity, _) =>
-        ignore(entity).map(_ =>
-          throw UnexpectedStatusCode(createSessionUri, other))
+        ignore(entity).map(
+          _ => throw UnexpectedStatusCode(createSessionUri, other)
+        )
     }
   }
 
@@ -295,10 +342,11 @@ final class ConsulCoordination(
   }
 
   private def send(baseRequest: HttpRequest) = {
-    def request: HttpRequest = settings.accessToken.map { token =>
-      val newHeaders = baseRequest.headers :+ HttpHeaders.ConsulToken(token)
-      baseRequest.copy(headers = newHeaders)
-    } getOrElse baseRequest
+    def request: HttpRequest =
+      settings.accessToken.map { token =>
+        val newHeaders = baseRequest.headers :+ ConsulToken(token)
+        baseRequest.copy(headers = newHeaders)
+      } getOrElse baseRequest
 
     Source
       .single(request)
